Add a directional shader for light rendering
This allows to render light as a cone in a certain direction with an inner
circle that is illuminating all directions. It is basically implemented as a
mask when rendering into the light texture. A separate fragment shader for the
first of the three render passes is used to set the light intensity based on
where the fragment is relative to the light source.

Parameters:
  * direction of the light
  * opening angle of inner cone (full intensity)
  * opening angle of outer cone (intensity fading out)
  * radius of inner circle around light center (full intensity within circle)
  * radius of outer circle around light center (intensity fading out)

The parameters are hard-coded at the moment but will be made script-accessible
in a subsequent commit.